Add TagContext and integrate TagProvider into App component
